Metro Made Simple Final Project
This project has a total of ten modules. Booking a ticket, Timings of metro, Recharge my metro card, Lost my metro card, Check my balance, Show tickets, Purchase a gift card, Transaction History, and Exit. A few of these modules are just an improved version of the already existing systems. On the contrary, some modules are this team's original ideas, such as purchasing a gift card. 

The objectives include giving a clear picture to people on how to use metro more efficiently and making the process of using metro easier. This project also attempts to make the entire metro system more user friendly. 

Instructions to run:
First, the project will be asking the user the username and password. You will have to enter the hardcoded details (Username - 123, Password - 1234) in order to start the modules. Depending on the module number, enter a number to take you to that module. The rest of the instructions are each different, according to the module. They are all given in the output of the code itself. For example, to book tickets, the output will hold the questions, such as asking for the destinations and the number of seats.

Please go through this link. Our Team, PyFive had worked on a video to explain Metro Made Simple.
https://www.youtube.com/watch?v=os70VntTOpY

Thank you,
PyFive